refactor: 마크 하이라이팅하는 커스텀훅 따로 분리
