+++ src/components/contentDetail/ScriptWithTime.tsx
@@ -3,11 +3,8 @@ import useMarkEvent from '@/hooks/contentDetail/useMarkEvent';
 import { tooltipHandlers } from '@/handlers/contentDetail/handleToolTip';
 import { ScriptNTimeProps } from '@/types/interface';
 import Tooltip from './Tooltip';
-import { useLayoutEffect } from 'react';
-import useScrappedStore from '@/store/useScrappedStore';
 import useMatchedStore from '@/store/useMatchedStore';
 import useToastStore from '@/store/useToastStore';
-import { useScrollToElement } from '@/hooks/contentDetail/useScrollToElement';
 
 export default function ScriptWithTime({
   id,
@@ -24,37 +21,14 @@ export default function ScriptWithTime({
     isScrapped: boolean;
     index: number | null;
   }>({ content: '', term: '', x: 0, y: 0, isScrapped: false, index: null });
-  const scrappedData = useScrappedStore((state) => state.scrappedData);
   const matched = useMatchedStore((state) => state.matched);
   const setMatched = useMatchedStore((state) => state.setMatched);
   const hideToast = useToastStore((state) => state.hideToast);
-
-  const { elementRefs } = useScrollToElement();
 
   useMarkEvent((event) => {
     hideToast();
     tooltipHandlers.handleClick(event, setTooltip, dictionaries, setMatched);
   });
-
-  useLayoutEffect(() => {
-    const marks = document.querySelectorAll('mark');
-    marks.forEach((mark, index) => {
-      const dict = scrappedData[index];
-      if (dict?.scrapped) {
-        mark.style.backgroundColor = '#A7FFB4';
-        mark.style.padding = '2px 1px';
-        mark.style.cursor = 'pointer';
-        mark.className = `mark mark-${dict?.dictionaryId}`;
-      } else {
-        mark.style.backgroundColor = '#FFED85';
-        mark.style.padding = '2px 1px';
-        mark.style.cursor = 'pointer';
-        mark.className = `mark mark-${dict?.dictionaryId}`;
-      }
-
-      elementRefs.current[mark.className] = mark as HTMLElement;
-    });
-  }, [scrappedData, elementRefs]);
 
   return (
     <div
